added some forgotten jsDocs + split booking into two functions
