+++ src/Components/booking.tsx
@@ -1,38 +1,16 @@
+import { useState } from 'react';
 import DatePicker from 'react-datepicker';
 import 'react-datepicker/dist/react-datepicker.css';
-import { useState } from 'react';
 import { userInfo } from '../utilities/localstorage';
 
-import { bookVenue } from '../API/booking/book';
-
-interface Booking {
-  dateFrom: string;
-  dateTo: string;
-}
+import { Booking } from '../Types/common';
+import { handleSubmitBooking } from '../UI/venue/booking';
 
 interface BookingFormProps {
   maxGuests: number;
   bookings: Booking[];
   id: string;
 }
-
-/**
- * BookingForm component allows users to select check-in and check-out dates,
- * choose the number of guests, and submit the booking for a venue.
- * It validates the booking dates to ensure there are no conflicts with existing bookings.
- *
- * @component
- * @example
- * const bookings = [{ dateFrom: '2025-05-15', dateTo: '2025-05-20' }];
- * return <BookingForm maxGuests={4} bookings={bookings} id="venue123" />;
- *
- * @param {Object} props - Component props.
- * @param {number} props.maxGuests - The maximum number of guests allowed for the venue.
- * @param {Array} props.bookings - A list of existing bookings, with start and end dates.
- * @param {string} props.id - The ID of the venue being booked.
- *
- * @returns {JSX.Element} A form component that allows users to book a venue.
- */
 
 export const BookingForm = ({ maxGuests, bookings, id }: BookingFormProps) => {
   const [checkIn, setCheckIn] = useState<Date | null>(null);
@@ -48,24 +26,7 @@ export const BookingForm = ({ maxGuests, bookings, id }: BookingFormProps) => {
   const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
 
-    if (!checkIn || !checkOut) return alert('Select both dates');
-    if (checkIn >= checkOut) return alert('Checkout must be after check-in');
-
-    const hasConflict = bookings.some((booking) => {
-      const existingStart = new Date(booking.dateFrom);
-      const existingEnd = new Date(booking.dateTo);
-
-      return checkIn < existingEnd && checkOut > existingStart;
-    });
-    if (hasConflict) {
-      document.getElementById('bookingErrorDates')?.classList.remove('hidden');
-      setTimeout(() => {
-        document.getElementById('bookingErrorDates')?.classList.add('hidden');
-      }, 5000);
-      return;
-    }
-
-    await bookVenue({ checkIn, checkOut, guests, venueId: id });
+    await handleSubmitBooking(checkIn, checkOut, guests, id, bookings); // Use the new function
   };
 
   return (
